backend/requirements.txt
# FastAPI & server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database & ORM
sqlmodel>=0.0.16

# Config
pydantic-settings>=2.6.0

# Deep Agent (LangChain-based)
deepagents>=0.4.0
langchain>=0.3.0
langgraph>=0.2.0
# LLM providers (默认 DeepSeek，可选 OpenAI/Anthropic)
langchain-deepseek>=1.0.0
langchain-openai>=0.2.0

# HTTP client for tool use
httpx>=0.27.0

# PDF 读取（read_pdf 工具）
pypdf>=5.0.0
pdfplumber>=0.11.0

# 技能 manifest 解析
pyyaml>=6.0
